🔥 Remove directory and move util fn
