--- static/app/components/events/contextSummary/contextSummaryNoSummary.tsx
@@ -1,15 +1,0 @@
-import Item from './item';
-
-type Props = {
-  title: string;
-};
-
-function ContextSummaryNoSummary({title}: Props) {
-  return (
-    <Item>
-      <h3 data-test-id="no-summary-title">{title}</h3>
-    </Item>
-  );
-}
-
-export default ContextSummaryNoSummary;
