--- static/app/components/events/contextSummary/types.tsx
@@ -1,23 +1,0 @@
-import type {Event} from 'sentry/types/event';
-
-type Meta = NonNullable<Event['_meta']>;
-
-/**
- * The props passed to each context item
- */
-export type ContextItemProps<D, M extends keyof Meta> = {
-  data: D;
-  meta: Meta[M];
-  omitUnknownVersion?: boolean;
-  unknownTitle?: string;
-};
-
-/**
- * Defines a context type
- */
-export type Context = {
-  Component: React.ComponentType<ContextItemProps<any, any>>;
-  keys: string[];
-  omitUnknownVersion?: boolean;
-  unknownTitle?: string;
-};
