+++ static/app/components/events/contexts/utils.tsx
@@ -5,7 +5,6 @@ import moment from 'moment-timezone';
 
 import UserAvatar from 'sentry/components/avatar/userAvatar';
 import ContextIcon from 'sentry/components/events/contexts/contextIcon';
-import {generateIconName} from 'sentry/components/events/contextSummary/utils';
 import {removeFilterMaskedEntries} from 'sentry/components/events/interfaces/utils';
 import StructuredEventData from 'sentry/components/structuredEventData';
 import {t} from 'sentry/locale';
@@ -111,6 +110,57 @@ const CONTEXT_TYPES = {
   // we want to keep it here for now so it works for existing versions
   'Memory Info': MemoryInfoEventContext,
 };
+
+/**
+ * Generates the class name used for contexts
+ */
+export function generateIconName(
+  name?: string | boolean | null,
+  version?: string
+): string {
+  if (!defined(name) || typeof name === 'boolean') {
+    return '';
+  }
+
+  const lowerCaseName = name.toLowerCase();
+
+  // amazon fire tv device id changes with version: AFTT, AFTN, AFTS, AFTA, AFTVA (alexa), ...
+  if (lowerCaseName.startsWith('aft')) {
+    return 'amazon';
+  }
+
+  if (lowerCaseName.startsWith('sm-') || lowerCaseName.startsWith('st-')) {
+    return 'samsung';
+  }
+
+  if (lowerCaseName.startsWith('moto')) {
+    return 'motorola';
+  }
+
+  if (lowerCaseName.startsWith('pixel')) {
+    return 'google';
+  }
+
+  const formattedName = name
+    .split(/\d/)[0]
+    .toLowerCase()
+    .replace(/[^a-z0-9\-]+/g, '-')
+    .replace(/\-+$/, '')
+    .replace(/^\-+/, '');
+
+  if (formattedName === 'edge' && version) {
+    const majorVersion = version.split('.')[0];
+    const isLegacyEdge = majorVersion >= '12' && majorVersion <= '18';
+
+    return isLegacyEdge ? 'legacy-edge' : 'edge';
+  }
+
+  if (formattedName.endsWith('-mobile')) {
+    return formattedName.split('-')[0];
+  }
+
+  return formattedName;
+}
 
 export function getContextComponent(type: string) {
   return CONTEXT_TYPES[type] || plugins.contexts[type] || CONTEXT_TYPES.default;
